requirements.txt
numpy==1.23.2
Pillow==9.2.0
pyfiglet==0.8.post1